Adjustment to allow changing breed name
